move zip retrieval retries into pkg/store
